pyproject.toml
[tool.poetry]
name = "midrasai"
version = "0.1.4"
description = ""
authors = ["ajac-zero <ajcardoza2000@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
midrasai = "midrasai.cli:midrascli"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1"
qdrant-client = "^1.11.1"
pdf2image = "^1.17.0"
torch = { version = "^2.4", optional = true }
transformers = { version = "^4.44", optional = true }
peft = { version = "^0.12", optional = true }
pillow = { version = "^10", optional = true }
fastapi = { extras = ["standard"], version = "^0", optional = true }
huggingface-hub = { extras = ["cli"], version = "^0", optional = true }

[tool.poetry.extras]
local = [
    "torch",
    "transformers",
    "peft",
    "pillow",
    "fastapi",
    "huggingface-hub",
]

[tool.poetry.group.local.dependencies]
torch = "^2.4"
transformers = "^4.44"
peft = "^0.12"
pillow = "^10.4"
fastapi = { extras = ["standard"], version = "^0" }
huggingface-hub = { extras = ["cli"], version = "^0" }


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.3"
pytest = "^8.3.2"


[tool.poetry.group.docs.dependencies]
ipykernel = "^6.29.5"
matplotlib = "^3.9.2"
google-generativeai = "^0.7.2"

[tool.ruff]
fix = true

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
fixable = ["ALL"]
exclude = ["examples"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
